Created chickenpox by sex
It would be interesting to see if there is any evidence of a link between vaccine effectiveness and sex of the child. Calculated the ratio of the number of children who contracted chickenpox but were vaccinated against it (at least one varicella dose) versus those who were vaccinated but did not contract chicken pox. The function returns results by sex.