Created corr-chickenpox
If we wanted to know if vaccines work, we might look at the correlation between the use of the vaccine and whether it results in prevention of the infection or disease. In this question, we see if there is a correlation between having had the chicken pox and the number of chickenpox vaccine doses given (varicella).